requirements.txt
PyQt6==6.6.1
PyQt6-WebEngine==6.6.0
pyinstaller==6.3.0
requests==2.31.0
retry==0.9.2
fake-useragent==1.0.1
dotmap
Telethon==1.28.5
fire==0.5.0
aiohttp
PySocks==1.7.1
python-socks==2.3.0
PyYAML==6.0.1
colorlog==6.7.0
PySocks==1.7.1
opencv-python
Pillow==9.5.0
numpy==1.21.6
selenium==4.11.2
asyncer==0.0.2
croniter==1.4.1